fix: update example to newest material  ui
