+++ examples/demo/package.json
@@ -26,7 +26,11 @@
   ],
   "dependencies": {
     "@berviantoleo/react-multi-crop": "^0.20.36",
-    "@material-ui/core": "^4.12.3",
+    "@emotion/react": "^11.7.0",
+    "@emotion/styled": "^11.6.0",
+    "@mui/icons-material": "^5.2.0",
+    "@mui/material": "^5.2.1",
+    "@mui/styles": "^5.2.1",
     "@types/jest": "^27.0.3",
     "@types/node": "^16.11.10",
     "@types/react": "^17.0.37",
