+++ examples/demo/src/components/ActionButton.tsx
@@ -1,6 +1,7 @@
 import React from "react";
-import Grid from "@material-ui/core/Grid";
-import Button from "@material-ui/core/Button";
+import Button from "@mui/material/Button";
+import Grid from "@mui/material/Grid";
+
 import { ReactMultiCrop, IOutputData } from "@berviantoleo/react-multi-crop";
 
 export default function ActionButton(): JSX.Element {
