+++ examples/demo/src/context/ColorModeContext.tsx
@@ -1,0 +1,6 @@
+import React from "react";
+
+export const ColorModeContext = React.createContext({
+  // eslint-disable-next-line
+  toggleColorMode: () => {},
+});
